requirements.txt
confluent-kafka==1.7.0
python-dateutil==2.8.2
scikit-learn
numpy
# scipy
pandas
matplotlib
wfdb
# pyspark
# pyspark[sql]
# pyhealth
pyarrow
mysql-connector-python==8.0.32